Revert "Dev (#2216)" (#2217)
This reverts commit 39993abf9e3231b7aadae0fdaa4ef75675dda541.

+++ packages/app/components/settings/clear-cache-btn.tsx
@@ -1,18 +1,28 @@
 import { useCallback } from "react";
 
+import { useAlert } from "@showtime-xyz/universal.alert";
 import { Zap } from "@showtime-xyz/universal.icon";
 
 import { deleteAppCache } from "app/lib/delete-cache";
 
-import { toast } from "design-system/toast";
-
 import { AccountSettingItem } from "./settings-account-item";
 
 export const SettingClearAppCache = () => {
+  const Alert = useAlert();
   const clearAppCache = useCallback(() => {
-    deleteAppCache();
-    toast.success("Cleared!");
-  }, []);
+    Alert.alert("Clear app cache?", "", [
+      {
+        text: "Confirm",
+        onPress: async () => {
+          deleteAppCache();
+        },
+        style: "destructive",
+      },
+      {
+        text: "Cancel",
+      },
+    ]);
+  }, [Alert]);
   return (
     <AccountSettingItem
       title="Clear App cache"
